+++ code/haitiwater/apps/dashboard/views.py
@@ -4,6 +4,7 @@ from ..water_network.models import Element
 from ..consumers.models import Consumer
 from django.template.loader import render_to_string
 from haitiwater.settings import PROJECT_VERSION, PROJECT_NAME
+from ..utils.get_data import *
 
 
 def index(request):
@@ -22,36 +23,3 @@ def index(request):
         context['amount_individual_consumers'] = get_amount_indiv_consummer(request.user.profile.zone)
     return HttpResponse(template.render(context, request))
 
-#TODO zone
-def get_amount(type, zone):
-    result = 0
-    for elem in Element.objects.filter(type=type):
-        if elem.zone.name in zone.subzones:
-            result += 1
-    return result
-
-def get_amount_fountain(zone):
-    return get_amount("FOUNTAIN", zone)
-
-def get_amount_kiosk(zone):
-    return get_amount("KIOSK", zone)
-
-def get_amount_individual(zone):
-    return get_amount("INDIVIDUAL", zone)
-
-def get_amount_pipe(zone):
-    return get_amount("PIPE", zone)
-
-def get_amount_consumer(zone):
-    res = 0
-    for consumer in Consumer.objects.all():
-        if consumer.water_outlet.zone.name in zone.subzones:
-            res += 1
-    return res
-
-def get_amount_indiv_consummer(zone):
-    result = 0
-    for consumer in Consumer.objects.all():
-        if consumer.water_outlet.zone.name in zone.subzones:
-            result += consumer.household_size
-    return result
