+++ code/haitiwater/apps/report/views.py
@@ -2,6 +2,7 @@ from django.http import HttpResponse
 from django.template import loader
 from ..water_network.models import Element
 from ..report.models import Report
+from ..utils.get_data import *
 from haitiwater.settings import PROJECT_VERSION, PROJECT_NAME
 
 
@@ -11,16 +12,6 @@ def index(request):
         'project_version': PROJECT_VERSION,
         'project_name': PROJECT_NAME,
         'current_period': 'Septembre',  # Todo Backend
-        'water_outlets': get_outlets()
+        'water_outlets': get_outlets(request)
     }
     return HttpResponse(template.render(context, request))
-
-def get_outlets():
-    all_outlets = Element.objects.all()
-    result = []
-    for elem in all_outlets:
-        if elem.type in ["KIOSK", "FOUNTAIN", "INDIVIDUAL"]:
-            reports = Report.objects.filter(water_outlet=elem.id) #TODO add filter by month
-            if len(reports) == 0:
-                result.append((elem.id, elem.name))
-    return result
